Код управления и слежки за временем вынесен из CRenderDevice в CTimeManager

+++ Sources/xrEngine/R_Backend.h
@@ -497,24 +497,24 @@ class ENGINE_API CBackend
 	}
 
 	void u_compute_texgen_screen(Fmatrix& dest);
-	void set_viewport_geometry(float w, float h, ref_geom geometry, u32& vOffset);
+	void set_viewport_geometry(u32 w, u32 h, ref_geom geometry, u32& vOffset);
 
 	void set_Render_Target_Surface(const ref_rt& _1, const ref_rt& _2 = NULL, const ref_rt& _3 = NULL, const ref_rt& _4 = NULL);
 	void set_Render_Target_Surface(u32 W, u32 H, IDirect3DSurface9* _1, IDirect3DSurface9* _2 = NULL, IDirect3DSurface9* _3 = NULL, IDirect3DSurface9* _4 = NULL);
 	void set_Depth_Buffer(IDirect3DSurface9* zb);
 	void clear_Depth_Buffer(IDirect3DSurface9* zb);
 
-	void set_viewport_geometry(float w, float h, u32& vOffset);
+	void set_viewport_geometry(u32 w, u32 h, u32& vOffset);
 	void set_viewport_geometry(ref_geom geometry, u32& vOffset);
 	void set_viewport_geometry(u32& vOffset);
 
-	void render_viewport_geometry(float w, float h);
+	void render_viewport_geometry(u32 w, u32 h);
 
 	void RenderViewportSurface();
-	void RenderViewportSurface(float w, float h, IDirect3DSurface9* _1, IDirect3DSurface9* zb = NULL);
+	void RenderViewportSurface(u32 w, u32 h, IDirect3DSurface9* _1, IDirect3DSurface9* zb = NULL);
 	void RenderViewportSurface(IDirect3DSurface9* _1);
 	void RenderViewportSurface(const ref_rt& _1, IDirect3DSurface9* zb = NULL);
-	void RenderViewportSurface(float w, float h, const ref_rt& _1, const ref_rt& _2 = NULL, const ref_rt& _3 = NULL, const ref_rt& _4 = NULL);
+	void RenderViewportSurface(u32 w, u32 h, const ref_rt& _1, const ref_rt& _2 = NULL, const ref_rt& _3 = NULL, const ref_rt& _4 = NULL);
 
 
 	void RenderToMipLevel(ref_rt target, u32 mip_level);
